Tests/LibWeb/Text/expected/UIEvents/mouse-boundary-events-offset.txt

> enter inner
pointerover target=inner current=inner related=body offset=(1, 3)
pointerover target=inner current=middle related=body offset=(1, 3)
pointerover target=inner current=outer related=body offset=(1, 3)
mouseover target=inner current=inner related=body offset=(1, 3)
mouseover target=inner current=middle related=body offset=(1, 3)
mouseover target=inner current=outer related=body offset=(1, 3)
pointerenter target=outer current=outer related=body offset=(35, 45)
mouseenter target=outer current=outer related=body offset=(35, 45)
pointerenter target=middle current=middle related=body offset=(15, 20)
mouseenter target=middle current=middle related=body offset=(15, 20)
pointerenter target=inner current=inner related=body offset=(1, 3)
mouseenter target=inner current=inner related=body offset=(1, 3)
> leave inner and middle, remain in outer
pointerout target=inner current=inner related=outer offset=(-34, -37)
pointerout target=inner current=middle related=outer offset=(-34, -37)
pointerout target=inner current=outer related=outer offset=(-34, -37)
mouseout target=inner current=inner related=outer offset=(-34, -37)
mouseout target=inner current=middle related=outer offset=(-34, -37)
mouseout target=inner current=outer related=outer offset=(-34, -37)
pointerleave target=inner current=inner related=outer offset=(-34, -37)
pointerleave target=middle current=middle related=outer offset=(-20, -20)
mouseleave target=inner current=inner related=outer offset=(-34, -37)
mouseleave target=middle current=middle related=outer offset=(-20, -20)
pointerover target=outer current=outer related=inner offset=(0, 5)
mouseover target=outer current=outer related=inner offset=(0, 5)
> enter inner again
pointerout target=outer current=outer related=inner offset=(35, 45)
mouseout target=outer current=outer related=inner offset=(35, 45)
pointerover target=inner current=inner related=outer offset=(1, 3)
pointerover target=inner current=middle related=outer offset=(1, 3)
pointerover target=inner current=outer related=outer offset=(1, 3)
mouseover target=inner current=inner related=outer offset=(1, 3)
mouseover target=inner current=middle related=outer offset=(1, 3)
mouseover target=inner current=outer related=outer offset=(1, 3)
pointerenter target=middle current=middle related=outer offset=(15, 20)
mouseenter target=middle current=middle related=outer offset=(15, 20)
pointerenter target=inner current=inner related=outer offset=(1, 3)
mouseenter target=inner current=inner related=outer offset=(1, 3)
> leave all
pointerout target=inner current=inner related=out offset=(131, 113)
pointerout target=inner current=middle related=out offset=(131, 113)
pointerout target=inner current=outer related=out offset=(131, 113)
mouseout target=inner current=inner related=out offset=(131, 113)
mouseout target=inner current=middle related=out offset=(131, 113)
mouseout target=inner current=outer related=out offset=(131, 113)
pointerleave target=inner current=inner related=out offset=(131, 113)
pointerleave target=middle current=middle related=out offset=(145, 130)
pointerleave target=outer current=outer related=out offset=(165, 155)
mouseleave target=inner current=inner related=out offset=(131, 113)
mouseleave target=middle current=middle related=out offset=(145, 130)
mouseleave target=outer current=outer related=out offset=(165, 155)
